evals moved + readme
