+++ environments/eval_environments/vision_evals/chartqa_environment.py
@@ -8,10 +8,10 @@ from pathlib import Path
 from typing import List, Optional, Tuple
 
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 
 class ChartQA(EvalBase):
@@ -142,17 +142,11 @@ Question: {query}"""
         # Non-numeric: exact match (case-insensitive)
         return pred.lower() == ans.lower()
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -189,10 +183,6 @@ Question: {query}"""
 if __name__ == "__main__":
     asyncio.run(
         eval_runner(
-            ChartQA,
-            subset="human",
-            relaxed_tolerance=0.05,
-            temperature=0.0,
-            max_tokens=2048,
+            ChartQA(subset="human", relaxed_tolerance=0.05, temperature=0.0, max_tokens=2048)
         )
     )
