+++ environments/eval_environments/vision_evals/mathverse_environment.py
@@ -9,10 +9,10 @@ from typing import List, Optional, Tuple
 
 import openai
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 EXTRACT_ICL_EXAMPLES = [
     "1.\nModel response: 'The perimeter of the sector is approximately (-2, 1)'\n"
@@ -245,17 +245,11 @@ Judgement:"""
 
         return False
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -328,12 +322,13 @@ Judgement:"""
 if __name__ == "__main__":
     asyncio.run(
         eval_runner(
-            MathVerse,
-            split="testmini",
-            use_cot=False,
-            use_gpt_evaluation=True,
-            judge_model="gpt-4o-mini",
-            temperature=0.0,
-            max_tokens=2048,
+            MathVerse(
+                split="testmini",
+                use_cot=False,
+                use_gpt_evaluation=True,
+                judge_model="gpt-4o-mini",
+                temperature=0.0,
+                max_tokens=2048,
+            )
         )
     )
