+++ environments/eval_environments/vision_evals/mathvision_environment.py
@@ -9,10 +9,10 @@ from typing import Dict, List, Optional, Tuple
 
 import openai
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 ICL_EXAMPLES = [
     """Hint: Please answer the question and provide the final answer at the end.
@@ -263,17 +263,11 @@ Then extract the answer from the model response and type it at the end of the pr
 
         return is_equal(prediction, answer)
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -335,11 +329,12 @@ Then extract the answer from the model response and type it at the end of the pr
 if __name__ == "__main__":
     asyncio.run(
         eval_runner(
-            MathVision,
-            split="testmini",
-            use_gpt_extraction=True,
-            judge_model="gpt-4o-mini",
-            temperature=0.0,
-            max_tokens=2048,
+            MathVision(
+                split="testmini",
+                use_gpt_extraction=True,
+                judge_model="gpt-4o-mini",
+                temperature=0.0,
+                max_tokens=2048,
+            )
         )
     )
