+++ environments/eval_environments/vision_evals/mathvista_environment.py
@@ -9,10 +9,10 @@ from typing import Dict, List, Optional, Tuple
 
 import openai
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 ICL_EXAMPLES = [
     """
@@ -322,17 +322,11 @@ class MathVista(EvalBase):
 
         return pred.lower() == ans.lower()
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -400,12 +394,13 @@ class MathVista(EvalBase):
 if __name__ == "__main__":
     asyncio.run(
         eval_runner(
-            MathVista,
-            split="testmini",
-            use_query=True,
-            use_gpt_extraction=True,
-            judge_model="gpt-4o-mini",
-            temperature=0.0,
-            max_tokens=4096,
+            MathVista(
+                split="testmini",
+                use_query=True,
+                use_gpt_extraction=True,
+                judge_model="gpt-4o-mini",
+                temperature=0.0,
+                max_tokens=4096,
+            )
         )
     )
