+++ environments/eval_environments/vision_evals/mmstar_environment.py
@@ -1,4 +1,4 @@
-"""BLINK evaluation environment."""
+"""MMStar evaluation environment."""
 
 import asyncio
 import base64
@@ -7,78 +7,53 @@ from string import ascii_uppercase
 from typing import List, Optional, Tuple
 
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 from environments.eval_environments.eval_helpers import (
     extract_letter_from_answer_tag,
     extract_mcqa_answer_with_fallback,
 )
 
 
-class BLINK(EvalBase):
-    """BLINK evaluation - visual perception benchmark."""
+class MMStar(EvalBase):
+    """MMStar evaluation - expert-level multimodal benchmark."""
 
     def setup_data(self) -> list:
         split = getattr(self, "split", "val")
-        task = getattr(self, "task", "Counting")  # One of the BLINK task categories
 
         try:
-            dataset = load_dataset("BLINK-Benchmark/BLINK", task, split=split)
-            print(f"Loaded {len(dataset)} examples from BLINK ({split}, {task})")
+            dataset = load_dataset("Lin-Chen/MMStar", split=split)
+            print(f"Loaded {len(dataset)} examples from MMStar ({split})")
             return list(dataset)
         except Exception as e:
-            print(f"Warning: Could not load BLINK: {e}")
+            print(f"Warning: Could not load MMStar: {e}")
             try:
-                tasks = [
-                    "Counting",
-                    "Spatial_Relation",
-                    "Object_Localization",
-                    "Visual_Similarity",
-                ]
-                all_data = []
-                for t in tasks:
-                    try:
-                        ds = load_dataset("BLINK-Benchmark/BLINK", t, split=split)
-                        for item in ds:
-                            item["task"] = t
-                            all_data.append(item)
-                    except Exception:
-                        pass
-                if all_data:
-                    print(f"Loaded {len(all_data)} examples from BLINK ({split})")
-                    return all_data
-                raise ValueError(f"Could not load BLINK dataset: {e}")
+                dataset = load_dataset("lmms-lab/MMStar", split=split)
+                print(f"Loaded {len(dataset)} examples from MMStar ({split})")
+                return list(dataset)
             except Exception:
-                raise ValueError(f"Could not load BLINK dataset: {e}")
+                raise ValueError(f"Could not load MMStar dataset: {e}")
 
     def encode_image(self, pil_image: Image.Image) -> str:
         buffer = io.BytesIO()
         pil_image.save(buffer, format="PNG")
         return base64.b64encode(buffer.getvalue()).decode("utf-8")
 
-    def get_images(self, item: dict) -> List[str]:
-        """Get all images from item (BLINK can have multiple images)."""
-        images = []
-        for i in range(1, 5):
-            key = f"image_{i}"
+    def get_image_base64(self, item: dict) -> Optional[str]:
+        for key in ["image", "decoded_image"]:
             if key in item and item[key] is not None:
                 if isinstance(item[key], Image.Image):
-                    images.append(self.encode_image(item[key]))
-
-        if not images and "image" in item and item["image"] is not None:
-            if isinstance(item["image"], Image.Image):
-                images.append(self.encode_image(item["image"]))
-
-        return images
+                    return self.encode_image(item[key])
+        return None
 
     def build_messages(self, item: dict) -> List[dict]:
-        images = self.get_images(item)
+        image_base64 = self.get_image_base64(item)
         question = item.get("question", "")
 
         options = {}
-        for letter in ascii_uppercase[:6]:
+        for letter in ascii_uppercase[:6]:  # MMStar typically has up to 6 options
             if letter in item and item[letter] is not None:
                 val = item[letter]
                 if isinstance(val, str) and val.strip():
@@ -92,11 +67,11 @@ class BLINK(EvalBase):
             prompt += "\nPlease select the correct answer from the options above."
 
         content = []
-        for img_b64 in images:
+        if image_base64:
             content.append(
                 {
                     "type": "image_url",
-                    "image_url": {"url": f"data:image/png;base64,{img_b64}"},
+                    "image_url": {"url": f"data:image/png;base64,{image_base64}"},
                 }
             )
         content.append({"type": "text", "text": prompt})
@@ -115,17 +90,10 @@ class BLINK(EvalBase):
         letter, method = extract_mcqa_answer_with_fallback(response, num_choices)
         return letter, method
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
-
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -172,11 +140,4 @@ class BLINK(EvalBase):
 
 
 if __name__ == "__main__":
-    asyncio.run(
-        eval_runner(
-            BLINK,
-            split="val",
-            temperature=0.0,
-            max_tokens=256,
-        )
-    )
+    asyncio.run(eval_runner(MMStar(split="val", temperature=0.0, max_tokens=256)))
