+++ environments/eval_environments/vision_evals/mmt_bench_environment.py
@@ -1,4 +1,4 @@
-"""SEED-Bench2-Plus evaluation environment."""
+"""MMT-Bench evaluation environment."""
 
 import asyncio
 import base64
@@ -7,52 +7,53 @@ from string import ascii_uppercase
 from typing import List, Optional, Tuple
 
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 from environments.eval_environments.eval_helpers import (
     extract_letter_from_answer_tag,
     extract_mcqa_answer_with_fallback,
 )
 
 
-class SEEDBench2Plus(EvalBase):
-    """SEED-Bench2-Plus evaluation - comprehensive visual understanding benchmark."""
+class MMTBench(EvalBase):
+    """MMT-Bench evaluation - multi-task multimodal benchmark."""
 
     def setup_data(self) -> list:
-        split = getattr(self, "split", "test")
-        max_samples = getattr(self, "max_samples", None)
+        split = getattr(self, "split", "train")
+        max_samples = getattr(self, "max_samples", None)  # None = use all samples
 
         try:
-            # Use streaming to avoid memory issues with this large dataset
-            dataset = load_dataset("lmms-lab/SEED-Bench-2", split=split, streaming=True)
-
-            # Take samples from streaming dataset
+            # Try full dataset download first
+            dataset = load_dataset("OpenGVLab/MMT-Bench", split=split)
+            data = list(dataset)
             if max_samples:
-                data = list(dataset.take(max_samples))
-            else:
-                # Default to 1000 samples to avoid loading entire 24k dataset
-                data = list(dataset.take(1000))
-
-            print(f"Loaded {len(data)} examples from SEED-Bench2 ({split}, streaming)")
+                data = data[:max_samples]
+            print(f"Loaded {len(data)} examples from MMT-Bench ({split})")
             return data
         except Exception as e:
-            print(f"Warning: Could not load SEED-Bench2: {e}")
+            print(f"Warning: Full download failed, using streaming: {e}")
+            # Fallback to streaming if full download fails (known column mismatch issue)
             try:
                 dataset = load_dataset(
-                    "lmms-lab/SEED-Bench", split=split, streaming=True
+                    "OpenGVLab/MMT-Bench", split=split, streaming=True
                 )
                 if max_samples:
                     data = list(dataset.take(max_samples))
                 else:
-                    data = list(dataset.take(1000))
+                    # Stream all available samples
+                    data = []
+                    for i, item in enumerate(dataset):
+                        data.append(item)
+                        if i % 5000 == 0 and i > 0:
+                            print(f"  Streamed {i} samples...")
                 print(
-                    f"Loaded {len(data)} examples from SEED-Bench ({split}, streaming)"
+                    f"Loaded {len(data)} examples from MMT-Bench ({split}, streaming)"
                 )
                 return data
             except Exception:
-                raise ValueError(f"Could not load SEED-Bench2-Plus dataset: {e}")
+                raise ValueError(f"Could not load MMT-Bench dataset: {e}")
 
     def encode_image(self, pil_image: Image.Image) -> str:
         buffer = io.BytesIO()
@@ -65,40 +66,28 @@ class SEEDBench2Plus(EvalBase):
                 val = item[key]
                 if isinstance(val, Image.Image):
                     return self.encode_image(val)
-                elif isinstance(val, list) and len(val) > 0:
-                    # SEED-Bench-2 stores images as a list of PIL images
-                    if isinstance(val[0], Image.Image):
-                        return self.encode_image(val[0])
+                elif isinstance(val, str) and len(val) > 100:
+                    # Already base64-encoded string
+                    return val
         return None
 
     def build_messages(self, item: dict) -> List[dict]:
         image_base64 = self.get_image_base64(item)
         question = item.get("question", "")
+        hint = item.get("hint", "")
 
         options = {}
-        for letter in ascii_uppercase[:6]:
-            # Check for choice_a, choice_b format
-            choice_key = f"choice_{letter.lower()}"
-            if choice_key in item and item[choice_key] is not None:
-                val = item[choice_key]
-                if isinstance(val, str) and val.strip():
-                    options[letter] = val
-            elif letter in item and item[letter] is not None:
+        for letter in ascii_uppercase[:8]:  # Support up to 8 options
+            if letter in item and item[letter] is not None:
                 val = item[letter]
                 if isinstance(val, str) and val.strip():
                     options[letter] = val
 
-        if not options:
-            choices = item.get("choices", [])
-            if isinstance(choices, str):
-                try:
-                    choices = eval(choices)
-                except Exception:
-                    choices = []
-            for i, choice in enumerate(choices):
-                options[ascii_uppercase[i]] = choice
+        prompt = ""
+        if hint and str(hint).strip() and str(hint).lower() != "nan":
+            prompt += f"Hint: {hint}\n"
+        prompt += f"Question: {question}\n"
 
-        prompt = f"Question: {question}\n"
         if options:
             prompt += "Options:\n"
             for letter in sorted(options.keys()):
@@ -129,17 +118,11 @@ class SEEDBench2Plus(EvalBase):
         letter, method = extract_mcqa_answer_with_fallback(response, num_choices)
         return letter, method
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -152,21 +135,15 @@ class SEEDBench2Plus(EvalBase):
 
             answer = data_item.get("answer", "")
 
-            choices = data_item.get("choices", [])
-            if isinstance(choices, str):
-                try:
-                    choices = eval(choices)
-                except Exception:
-                    choices = []
-
-            num_choices = len(choices) if choices else 4
-            if num_choices == 0:
-                num_choices = sum(
-                    1
-                    for letter in ascii_uppercase[:6]
-                    if letter in data_item and data_item[letter] is not None
-                )
-                num_choices = max(num_choices, 4)
+            num_choices = sum(
+                1
+                for letter in ascii_uppercase[:8]
+                if letter in data_item
+                and data_item[letter] is not None
+                and isinstance(data_item[letter], str)
+                and data_item[letter].strip()
+            )
+            num_choices = max(num_choices, 4)
 
             extracted, method = self.extract_answer(response, num_choices)
 
@@ -175,11 +152,9 @@ class SEEDBench2Plus(EvalBase):
                 correct = extracted.upper() == str(answer).upper()
 
             sample = {
-                "id": data_item.get("index", data_item.get("question_id", "")),
+                "id": data_item.get("index", data_item.get("id", "")),
                 "question": data_item.get("question", "")[:200],
-                "category": data_item.get(
-                    "question_type_id", data_item.get("category", "")
-                ),
+                "task": data_item.get("task", ""),
                 "answer": answer,
                 "prediction": extracted,
                 "raw_response": response[:500],
@@ -195,10 +170,5 @@ class SEEDBench2Plus(EvalBase):
 
 if __name__ == "__main__":
     asyncio.run(
-        eval_runner(
-            SEEDBench2Plus,
-            split="test",
-            temperature=0.0,
-            max_tokens=256,
-        )
+        eval_runner(MMTBench(split="val", temperature=0.0, max_tokens=256))
     )
