+++ environments/eval_environments/vision_evals/realworldqa_environment.py
@@ -4,10 +4,10 @@ import io
 from typing import List, Tuple
 
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 
 class RealWorldQA(EvalBase):
@@ -77,17 +77,11 @@ Provide a brief, direct answer."""
 
         return False
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -122,11 +116,4 @@ Provide a brief, direct answer."""
 
 
 if __name__ == "__main__":
-    asyncio.run(
-        eval_runner(
-            RealWorldQA,
-            split="test",
-            temperature=0.0,
-            max_tokens=256,
-        )
-    )
+    asyncio.run(eval_runner(RealWorldQA(split="test", temperature=0.0, max_tokens=256)))
