+++ environments/eval_environments/vision_evals/vlmblind_environment.py
@@ -7,10 +7,10 @@ import re
 from typing import List, Optional, Tuple
 
 from datasets import load_dataset
-from openai import AsyncOpenAI
 from PIL import Image
 
-from environments.eval_environments.eval_base import EvalBase, eval_runner
+from atroposlib.envs.server_handling.server_manager import ServerManager
+from environments.eval_environments.eval import EvalBase, eval_runner
 
 
 class VLMBlind(EvalBase):
@@ -120,17 +120,11 @@ class VLMBlind(EvalBase):
         else:
             return answer_lower in response_lower, response_lower[:50]
 
-    async def run_item(self, client: AsyncOpenAI, data_item: dict) -> Tuple[dict, dict]:
+    async def run_item(self, server: ServerManager, data_item: dict) -> Tuple[dict, dict]:
         try:
             messages = self.build_messages(data_item)
 
-            gen_params = self.get_generation_params()
-            completion = await client.chat.completions.create(
-                model=self.model_name,
-                messages=messages,
-                temperature=gen_params["temperature"],
-                max_tokens=gen_params["max_tokens"],
-            )
+            completion = await self.chat_completion(server, messages)
 
             if not completion.choices:
                 return {"accuracy": 0.0}, {"error": "Empty response"}
@@ -167,10 +161,5 @@ class VLMBlind(EvalBase):
 
 if __name__ == "__main__":
     asyncio.run(
-        eval_runner(
-            VLMBlind,
-            split="test",
-            temperature=0.0,
-            max_tokens=512,
-        )
+        eval_runner(VLMBlind(split="test", temperature=0.0, max_tokens=512))
     )
